+++ kythe/go/serving/tools/kwazthis/kwazthis.go
@@ -200,6 +200,11 @@ func main() {
 		References:  true,
 		SourceText:  true,
 		DirtyBuffer: text,
+		Filter: []string{
+			schema.NodeKindFact,
+			schema.SubkindFact,
+			schema.AnchorLocFilter, // TODO(schroederc): remove once backwards-compatibility fix below is removed
+		},
 	})
 	if err != nil {
 		log.Fatal(err)
@@ -214,7 +219,13 @@ func main() {
 
 	en := json.NewEncoder(os.Stdout)
 	for _, ref := range decor.Reference {
-		start, end := parseAnchorSpan(nodes[ref.SourceTicket])
+		var start, end int
+		if ref.AnchorStart == nil || ref.AnchorEnd == nil {
+			// TODO(schroederc): remove this backwards-compatibility fix
+			start, end = parseAnchorSpan(nodes[ref.SourceTicket])
+		} else {
+			start, end = int(ref.AnchorStart.ByteOffset), int(ref.AnchorEnd.ByteOffset)
+		}
 
 		if start <= point && point < end {
 			var r reference